Update README.md
Sales and Taxes

Write a program that will input total sales that a business generates for a particular month. The program will also input the state and local sales tax percentage. It will output the total sales plus the state tax and local tax to be paid. The output should be in fixed notation with 2 decimal places.

Sample run:

Please input the total sales for the month
1080
Please input the state tax percentage in decimal form (.02 for 2%)
.06
Please input the local tax percentage in decimal form (.02 for 2%)
.02

The total sales for the month is $1080.00
The state tax for the month is $64.80
The local tax for the month is $21.60